Move lint merge command to tooling module
A monolithic CLI module is bad for the classpath, as the tools we want to use depend on different versions of unstable APIs.
