+++ redwood-tooling-lint/src/test/kotlin/app/cash/redwood/tooling/lint/ApiMergeCommandTest.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.redwood.cli
+package app.cash.redwood.tooling.lint
 
 import assertk.assertThat
 import assertk.assertions.isEqualTo
